% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxstats.R
\name{boxstats}
\alias{boxstats}
\title{Computes the statitics used to display the boxplot for a vector of values.}
\usage{
boxstats(y)
}
\arguments{
\item{y}{Column vector of size N x 1}
}
\value{
min : Minimum value in the vector
        q1 :  First quartile
        median : Median value
        q3 :  Third quartile
        max : Maximum value in the vector
        mean : Mean value
        std : Standard deviation
}
\description{
Computes the statitics used to display the boxplot for a vector of values.
}
\keyword{boxplot}

