% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{BoxStats_node}
\alias{BoxStats_node}
\title{Computes the statitics used to display the boxplot for a vector of values.}
\usage{
BoxStats_node(y)
}
\arguments{
\item{y}{Column vector of size N x 1}
}
\value{
min : Minimum value in the vector
        q1 :  First quartile
        median : Median value
        q3 :  Third quartile
        max : Maximum value in the vector
}
\description{
Computes the statitics used to display the boxplot for a vector of values.
}
\keyword{boxplot}

