% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/federation.R
\name{LRegress_Federation}
\alias{LRegress_Federation}
\title{Joins aggregates from two nodes to compute a linear regression. We have to compute beta from the linear equation:
y = A*beta + error, but data y is distributed over local nodes.}
\usage{
LRegress_Federation(beta1, Sigma1, beta2, Sigma2)
}
\arguments{
\item{beta1}{regression coefficients computed in nodes 1}

\item{Sigma1}{Covariance matrices of the regression coefficients in node 1.}

\item{beta2}{regression coefficients computed in nodes 2}

\item{Sigma2}{Covariance matrices of the regression coefficients in node 2.}
}
\value{
betaf  : Regression coefficient joining properly beta1,beta2, Sigma1 and Sigma2.
                 This will be the result the user will receive. The intermediate ones beta1 and beta2 are not shown.
        Sigmaf : Covariance matrix of joining properly the partial covariance matrices coming from different nodes.
}
\description{
Joins aggregates from two nodes to compute a linear regression. We have to compute beta from the linear equation:
y = A*beta + error, but data y is distributed over local nodes.
}
\keyword{regression}

