% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{LRegress_Node}
\alias{LRegress_Node}
\title{Computes the linear regression.  We have to compute beta from the linear equation: y = A*beta + error
This is the function that will be running at local node level. The data (input parameters: y, A) are obtained from the
local databases using a specific queries. These queries will be the same for all nodes.}
\usage{
LRegress_Node(y, A)
}
\arguments{
\item{y}{Data obtained from specific query to the local Database. y : column vector on size N x 1}

\item{A}{Design or regression matrix that will be obtained from specific query to the local Database. A : Matrix of size N x M , M : number of regressors}
}
\value{
betai  : regression coefficient computed in nodes 'i'.
        Sigmai : Covariance matrix of the regression coefficients betai.
}
\description{
Computes the linear regression.  We have to compute beta from the linear equation: y = A*beta + error
This is the function that will be running at local node level. The data (input parameters: y, A) are obtained from the
local databases using a specific queries. These queries will be the same for all nodes.
}
\keyword{regression}

