% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystats.R
\name{summarystats}
\alias{summarystats}
\title{Computes the summary statistics used to display the boxplot for a vector of values.}
\usage{
summarystats(y)
}
\arguments{
\item{y}{Column vector of size N x 1}
}
\value{
min : Minimum value in the vector
        q1 :  First quartile
        median : Median value
        q3 :  Third quartile
        max : Maximum value in the vector
        mean : Mean value
        std : Standard deviation
        sum: Sum of all values
        count: Number of values N
}
\description{
Computes the summary statistics used to display the boxplot for a vector of values.
}
\keyword{summarystats}

