% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystats_group.R
\name{summarystats_group}
\alias{summarystats_group}
\title{Computes the summary statistics used to display the summaryplot for a vector of values.}
\usage{
summarystats_group(partialStats)
}
\arguments{
\item{partialStats}{the results of summarystats function applied to multiple datasets and packed in one dataframe.}
}
\value{
min : Minimum value in the vector
        q1 :  First quartile
        median : Median value
        q3 :  Third quartile
        max : Maximum value in the vector
        mean : Mean value
        std : Standard deviation
        sum: Sum of all values
        count: Number of values N
}
\description{
Computes the summary statistics used to display the summaryplot for a vector of values.
}
\keyword{federation}
\keyword{summarystats}

