% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesummarystats.R
\name{tablesummarystats}
\alias{tablesummarystats}
\title{Computes the minimum, maximum, media, quantiles and other parameters to be used for plotting.
Only those variables that are numeric will be return with the previous statistics.}
\usage{
tablesummarystats(ytable, ycols2plot)
}
\arguments{
\item{ytable}{Data table type data.frame.}

\item{ycols2plot}{Names of the variables in the Table that will be plotted.}
}
\value{
fstats : Statistic necessary for plotting: minimum, maximum, media, quantiles. This is data.frame type.
}
\description{
Computes the minimum, maximum, media, quantiles and other parameters to be used for plotting.
Only those variables that are numeric will be return with the previous statistics.
}
\keyword{summarystats}
\keyword{table}
\keyword{tablesummarystats}

