% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesummarystats_group.R
\name{tablesummarystats_group}
\alias{tablesummarystats_group}
\title{Computes the statistics used to display the summaryplot for a vector of values.}
\usage{
tablesummarystats_group(listStats)
}
\arguments{
\item{listStats}{the results of tablesummarystat function applied to multiple datasets and packed in one list.}
}
\value{
min : Minimum value in the vector
        q1 :  First quartile
        median : Median value
        q3 :  Third quartile
        max : Maximum value in the vector
        mean : Mean value
        std : Standard deviation
        sum: Sum of all values
        count: Number of values N
}
\description{
Computes the statistics used to display the summaryplot for a vector of values.
}
\keyword{federation}
\keyword{summarystats}
\keyword{table}
\keyword{tablesummarystats}

